refactor: (Photo)/MarkerResponse class -> record로 변경
